_isDynatraceEnabled
