+++ lib/tracing/index.js
@@ -7,7 +7,21 @@ const { BatchSpanProcessor, SimpleSpanProcessor } = require('@opentelemetry/sdk-
 const { NodeTracerProvider } = require('@opentelemetry/sdk-trace-node')
 const { SemanticAttributes /* , SemanticResourceAttributes */ } = require('@opentelemetry/semantic-conventions')
 
-const { isDynatraceEnabled, _require } = require('../utils')
+const { _require } = require('../utils')
+
+function _isDynatraceEnabled() {
+  try {
+    const pkg = require(cds.root + '/package.json')
+    return Object.keys(pkg.dependencies).includes('@dynatrace/oneagent-sdk')
+  } catch (err) {
+    LOG._info &&
+      LOG.info(
+        'Unable to require package.json to check whether @dynatrace/oneagent-sdk is in dependencies due to error:',
+        err
+      )
+  }
+  return false
+}
 
 function _getSampler() {
   function _ignoreSpecifiedPaths(spanName, spanKind, attributes) {
@@ -99,7 +113,7 @@ module.exports = resource => {
   registerInstrumentations({ tracerProvider, instrumentations })
 
   // REVISIT: why not do if dynatrace is in deps?
-  if (process.env.NODE_ENV !== 'production' || !isDynatraceEnabled()) {
+  if (process.env.NODE_ENV !== 'production' || !_isDynatraceEnabled()) {
     const exporter = _getExporter()
     // REVISIT: why this difference?
     if (process.env.NODE_ENV !== 'production') {
