+++ lib/utils.js
@@ -65,20 +65,6 @@ function getResource() {
   return new Resource(attributes)
 }
 
-function isDynatraceEnabled() {
-  try {
-    const pkg = require(cds.root + '/package.json')
-    return Object.keys(pkg.dependencies).includes('@dynatrace/oneagent-sdk')
-  } catch (err) {
-    LOG._info &&
-      LOG.info(
-        'Unable to require package.json to check whether @dynatrace/oneagent-sdk is in dependencies due to error:',
-        err
-      )
-  }
-  return false
-}
-
 function _require(name) {
   name = Array.isArray(name) ? name[0] : name
   try {
@@ -92,6 +78,5 @@ function _require(name) {
 module.exports = {
   getDiagLogLevel,
   getResource,
-  isDynatraceEnabled,
   _require
 }
